Restructure pagination and formatting code. Moved the code for handling pagination to a separate package and the code for handling formatting to a separate file. Fixed inconsistency in pager struct names.
