+++ internals/secrethub/audit.go
@@ -1,15 +1,12 @@
 package secrethub
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"os"
-	"os/exec"
-	"strings"
-	"syscall"
+
+	"github.com/secrethub/secrethub-cli/internals/secrethub/pager"
 
 	"golang.org/x/crypto/ssh/terminal"
 
@@ -23,23 +20,19 @@ import (
 )
 
 var (
-	errPagerNotFound = errors.New("no terminal pager available. Please configure a terminal pager by setting the $PAGER environment variable or install \"less\" or \"more\"")
-	errPagerClosed   = errors.New("cannot write to closed terminal pager")
-	errNoSuchFormat  = func(format string) error { return errors.New("invalid format: " + format) }
+	errNoSuchFormat = func(format string) error { return errors.New("invalid format: " + format) }
 )
 
 const (
-	pagerEnvvar            = "$PAGER"
-	defaultTerminalWidth   = 80
-	fallbackPagerLineCount = 100
-	formatTable            = "table"
-	formatJSON             = "json"
+	defaultTerminalWidth = 80
+	formatTable          = "table"
+	formatJSON           = "json"
 )
 
 // AuditCommand is a command to audit a repo or a secret.
 type AuditCommand struct {
 	io                 ui.IO
-	newPaginatedWriter func(io.Writer) (pager, error)
+	newPaginatedWriter func(io.Writer) (io.WriteCloser, error)
 	path               api.Path
 	useTimestamps      bool
 	timeFormatter      TimeFormatter
@@ -53,7 +46,7 @@ type AuditCommand struct {
 func NewAuditCommand(io ui.IO, newClient newClientFunc) *AuditCommand {
 	return &AuditCommand{
 		io:                 io,
-		newPaginatedWriter: newPaginatedWriter,
+		newPaginatedWriter: pager.New,
 		newClient:          newClient,
 		terminalWidth: func(fd int) (int, error) {
 			w, _, err := terminal.GetSize(fd)
@@ -96,8 +89,8 @@ func (cmd *AuditCommand) run() error {
 	}
 
 	paginatedWriter, err := cmd.newPaginatedWriter(os.Stdout)
-	if err == errPagerNotFound {
-		paginatedWriter = newFallbackPaginatedWriter(os.Stdout)
+	if err == pager.ErrPagerNotFound {
+		paginatedWriter = pager.NewFallbackPager(os.Stdout)
 	} else if err != nil {
 		return err
 	}
@@ -130,197 +123,13 @@ func (cmd *AuditCommand) run() error {
 		}
 
 		err = formatter.Write(row)
-		if err == errPagerClosed {
+		if err == pager.ErrPagerClosed {
 			break
 		} else if err != nil {
 			return err
 		}
 	}
 	return nil
-}
-
-type listFormatter interface {
-	Write([]string) error
-}
-
-// newJSONFormatter returns a table formatter that formats the given table rows as json.
-func newJSONFormatter(writer io.Writer, fieldNames []string) *jsonFormatter {
-	for i := range fieldNames {
-		fieldNames[i] = toPascalCase(fieldNames[i])
-	}
-	return &jsonFormatter{
-		encoder: json.NewEncoder(writer),
-		fields:  fieldNames,
-	}
-}
-
-func toPascalCase(s string) string {
-	return strings.ReplaceAll(strings.Title(s), " ", "")
-}
-
-type jsonFormatter struct {
-	encoder *json.Encoder
-	fields  []string
-}
-
-// formatRow returns the json representation of the given row
-// with the configured field names as keys and the provided values
-func (f *jsonFormatter) Write(values []string) error {
-	if len(f.fields) != len(values) {
-		return fmt.Errorf("unexpected number of json fields")
-	}
-
-	jsonMap := make(map[string]string)
-	for i, element := range values {
-		jsonMap[f.fields[i]] = element
-	}
-
-	return f.encoder.Encode(jsonMap)
-}
-
-// newTableFormatter returns a list formatter that formats entries in a table.
-func newTableFormatter(writer io.Writer, tableWidth int, columns []tableColumn) *tableFormatter {
-	return &tableFormatter{
-		writer:     writer,
-		tableWidth: tableWidth,
-		columns:    columns,
-	}
-}
-
-type tableFormatter struct {
-	tableWidth           int
-	writer               io.Writer
-	computedColumnWidths []int
-	columns              []tableColumn
-	headerPrinted        bool
-}
-
-// Write writes the given values formatted in a table with the configured column widths and names.
-// The header of the table is printed on the first call, before any other value.
-func (f *tableFormatter) Write(values []string) error {
-	if !f.headerPrinted {
-		header := make([]string, len(f.columns))
-		for i, col := range f.columns {
-			header[i] = strings.ToUpper(col.name)
-		}
-		formattedHeader := f.formatRow(header)
-		_, err := f.writer.Write(formattedHeader)
-		if err != nil {
-			return err
-		}
-		f.headerPrinted = true
-	}
-
-	formattedRow := f.formatRow(values)
-	_, err := f.writer.Write(formattedRow)
-	return err
-}
-
-// formatRow formats the given table row to fit the configured width by
-// giving each cell an equal width and wrapping the text in cells that exceed it.
-func (f *tableFormatter) formatRow(row []string) []byte {
-	columnWidths := f.columnWidths()
-	grid := f.rowToGrid(row, columnWidths)
-
-	strRes := strings.Builder{}
-	for _, row := range grid {
-		strRes.WriteString(strings.Join(row, "  ") + "\n")
-	}
-	return []byte(strRes.String())
-}
-
-// rowToGrid returns a the given row split over a matrix in which all columns have equal length.
-// Longer values are split over multiple cells and shorter (or empty) ones are padded with " ".
-func (f *tableFormatter) rowToGrid(row []string, columnWidths []int) [][]string {
-	maxLinesPerCell := f.lineCount(row, columnWidths)
-
-	grid := make([][]string, maxLinesPerCell)
-	for i := 0; i < maxLinesPerCell; i++ {
-		grid[i] = make([]string, len(row))
-	}
-
-	for i, cell := range row {
-		columnWidth := columnWidths[i]
-		lineCount := len(cell) / columnWidth
-		for j := 0; j < lineCount; j++ {
-			begin := j * columnWidth
-			end := (j + 1) * columnWidth
-			grid[j][i] = cell[begin:end]
-		}
-
-		charactersLeft := len(cell) % columnWidth
-		if charactersLeft != 0 {
-			grid[lineCount][i] = cell[len(cell)-charactersLeft:] + strings.Repeat(" ", columnWidth-charactersLeft)
-		} else if lineCount < maxLinesPerCell {
-			grid[lineCount][i] = strings.Repeat(" ", columnWidth)
-		}
-
-		for j := lineCount + 1; j < maxLinesPerCell; j++ {
-			grid[j][i] = strings.Repeat(" ", columnWidth)
-		}
-	}
-
-	return grid
-}
-
-// lineCount returns the number of lines the given table row will occupy after splitting the
-// cell values that exceed their column width.
-func (f *tableFormatter) lineCount(row []string, widths []int) int {
-	maxLinesPerCell := 1
-	for i, value := range row {
-		lines := len(value) / widths[i]
-		if len(value)%widths[i] != 0 {
-			lines++
-		}
-		if lines > maxLinesPerCell {
-			maxLinesPerCell = lines
-		}
-	}
-	return maxLinesPerCell
-}
-
-// columnWidths returns the width of each column based on their maximum widths
-// and the table width.
-func (f *tableFormatter) columnWidths() []int {
-	if f.computedColumnWidths != nil {
-		return f.computedColumnWidths
-	}
-	res := make([]int, len(f.columns))
-
-	// Distribute the maximum width equally between all columns and repeatedly
-	// check if any of them have a smaller maximum width and can be shrunk.
-	// Stop when no columns can be further adjusted.
-	adjusted := true
-	columnsLeft := len(f.columns)
-	widthLeft := f.tableWidth - 2*(len(f.columns)-1)
-	widthPerColumn := widthLeft / columnsLeft
-	for adjusted {
-		adjusted = false
-		for i, col := range f.columns {
-			if res[i] == 0 && col.maxWidth != 0 && col.maxWidth < widthPerColumn {
-				res[i] = col.maxWidth
-				widthLeft -= col.maxWidth
-				columnsLeft--
-				adjusted = true
-			}
-		}
-		if columnsLeft == 0 {
-			for i := range res {
-				res[i] += widthLeft / len(res)
-			}
-			break
-		}
-		widthPerColumn = widthLeft / columnsLeft
-	}
-
-	// distribute the remaining width equally between columns with no maximum width.
-	for i := range res {
-		if res[i] == 0 {
-			res[i] = widthPerColumn
-		}
-	}
-	f.computedColumnWidths = res
-	return res
 }
 
 func (cmd *AuditCommand) iterAndAuditTable() (secrethub.AuditEventIterator, auditTable, error) {
@@ -363,150 +172,6 @@ func (cmd *AuditCommand) iterAndAuditTable() (secrethub.AuditEventIterator, audi
 	}
 
 	return nil, nil, ErrNoValidRepoOrSecretPath
-}
-
-// pager is an io.WriteCloser that returns errPagerClosed if it has been closed.
-type pager io.WriteCloser
-
-// newPaginatedWriter runs the terminal pager configured in the OS environment
-// and returns a writer that is piped to the standard input of the pager command.
-func newPaginatedWriter(outputWriter io.Writer) (pager, error) {
-	pager, err := pagerCommand()
-	if err != nil {
-		return nil, err
-	}
-
-	cmd := exec.Command(pager)
-
-	writer, err := cmd.StdinPipe()
-	if err != nil {
-		return nil, err
-	}
-
-	cmd.Stdout = outputWriter
-	cmd.Stderr = os.Stderr
-
-	err = cmd.Start()
-	if err != nil {
-		return nil, err
-	}
-	done := make(chan struct{}, 1)
-	go func() {
-		_ = cmd.Wait()
-		done <- struct{}{}
-	}()
-	return &paginatedWriter{writer: writer, cmd: cmd, done: done}, nil
-}
-
-// paginatedWriter is a writer that is piped to a terminal pager command.
-type paginatedWriter struct {
-	writer io.WriteCloser
-	cmd    *exec.Cmd
-	done   <-chan struct{}
-	closed bool
-}
-
-// Write pipes the data to the terminal pager.
-// It returns errPagerClosed if the terminal pager has been closed.
-func (p *paginatedWriter) Write(data []byte) (n int, err error) {
-	if p.isClosed() {
-		return 0, errPagerClosed
-	}
-	return p.writer.Write(data)
-}
-
-// Close closes the writer to the terminal pager and waits for the terminal pager to close.
-func (p *paginatedWriter) Close() error {
-	err := p.writer.Close()
-	if err != nil {
-		return err
-	}
-	if p.closed {
-		return nil
-	}
-	err = p.cmd.Process.Signal(syscall.SIGINT)
-	if err != nil {
-		err = p.cmd.Process.Kill()
-		if err != nil {
-			return err
-		}
-	}
-	<-p.done
-	return nil
-}
-
-// isClosed checks if the terminal pager process has been stopped.
-func (p *paginatedWriter) isClosed() bool {
-	if p.closed {
-		return true
-	}
-	select {
-	case <-p.done:
-		p.closed = true
-		return true
-	default:
-		return false
-	}
-}
-
-// pagerCommand returns the name of the terminal pager configured in the OS environment ($PAGER).
-// If no pager is configured it falls back to "less" than "more", returning an error if neither are available.
-func pagerCommand() (string, error) {
-	var pager string
-	var err error
-
-	pager, err = exec.LookPath(os.ExpandEnv(pagerEnvvar))
-	if err == nil {
-		return pager, nil
-	}
-
-	pager, err = exec.LookPath("less")
-	if err == nil {
-		return pager, nil
-	}
-
-	pager, err = exec.LookPath("more")
-	if err == nil {
-		return pager, nil
-	}
-
-	return "", errPagerNotFound
-}
-
-// newFallbackPaginatedWriter returns a pager that closes after outputting a fixed number of lines without pagination
-// and returns errPagerNotFound on the last (or any subsequent) write.
-func newFallbackPaginatedWriter(w io.WriteCloser) pager {
-	return &fallbackPager{
-		linesLeft: fallbackPagerLineCount,
-		writer:    w,
-	}
-}
-
-type fallbackPager struct {
-	writer    io.WriteCloser
-	linesLeft int
-}
-
-func (p *fallbackPager) Write(data []byte) (int, error) {
-	if p.linesLeft == 0 {
-		return 0, errPagerNotFound
-	}
-
-	lines := bytes.Count(data, []byte{'\n'})
-	if lines > p.linesLeft {
-		data = bytes.Join(bytes.Split(data, []byte{'\n'})[:p.linesLeft], []byte{'\n'})
-		data = append(data, '\n')
-	}
-	p.linesLeft -= bytes.Count(data, []byte{'\n'})
-	n, err := p.writer.Write(data)
-	if p.linesLeft == 0 {
-		err = errPagerNotFound
-	}
-	return n, err
-}
-
-func (p *fallbackPager) Close() error {
-	return p.writer.Close()
 }
 
 type tableColumn struct {
