+++ internals/secrethub/audit_secret_test.go
@@ -195,7 +195,7 @@ func TestAuditSecretCommand_run(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			// Setup
 			buffer := bytes.Buffer{}
-			tc.cmd.newPaginatedWriter = func(_ io.Writer) (pager, error) {
+			tc.cmd.newPaginatedWriter = func(_ io.Writer) (io.WriteCloser, error) {
 				return &fakes.Pager{Buffer: &buffer}, nil
 			}
 			// Act
