+++ internals/secrethub/formatter_test.go
@@ -1,11 +1,8 @@
 package secrethub
 
 import (
-	"bytes"
 	"strings"
 	"testing"
-
-	"github.com/secrethub/secrethub-cli/internals/secrethub/fakes"
 
 	"github.com/secrethub/secrethub-go/internals/assert"
 )
@@ -128,50 +125,6 @@ func Test_columnFormatter_formatRow(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			result := tc.formatter.formatRow(tc.row)
 			assert.Equal(t, string(result), tc.expected)
-		})
-	}
-}
-
-func TestFallbackPager_Write(t *testing.T) {
-	cases := map[string]struct {
-		pager       *fallbackPager
-		param       string
-		expected    string
-		expectedErr error
-	}{
-		"no lines left": {
-			pager:       &fallbackPager{linesLeft: 0},
-			expectedErr: errPagerNotFound,
-			param:       "test\n",
-			expected:    "",
-		},
-		"last line": {
-			pager:       &fallbackPager{linesLeft: 1},
-			expectedErr: errPagerNotFound,
-			param:       "test\n",
-			expected:    "test\n",
-		},
-		"print more": {
-			pager:       &fallbackPager{linesLeft: 2},
-			param:       "test1\ntest2\ntest3\ntest4",
-			expected:    "test1\ntest2\n",
-			expectedErr: errPagerNotFound,
-		},
-		"more lines left": {
-			pager:       &fallbackPager{linesLeft: 3},
-			expectedErr: nil,
-			param:       "test\n",
-			expected:    "test\n",
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			buffer := bytes.Buffer{}
-			tc.pager.writer = &fakes.Pager{Buffer: &buffer}
-			_, err := tc.pager.Write([]byte(tc.param))
-			assert.Equal(t, err, tc.expectedErr)
-			assert.Equal(t, buffer.String(), tc.expected)
 		})
 	}
 }
